refactor: input file move
