+++ apps/web/src/components/FeedBack/FeedBack.tsx
@@ -2,8 +2,9 @@
 
 import styled from 'styled-components';
 
+import Input from '@/components/Input';
+
 import { ISSUE_LABEL } from './FeedBack.constants';
-import Input from './Input';
 import Select from './Select';
 
 function FeedBack() {
